+++ src/main/java/com/lucenasoft/marketplace/models/ProductResponseDTO.java
@@ -1,6 +1,6 @@
 package com.lucenasoft.marketplace.models;
 
-public record ProductResponseDTO(Long id, String title, String imgURL, Integer price) {
+public record ProductResponseDTO(Long id, String title, String imgURL, Double price) {
     public ProductResponseDTO(ProductModel product){
         this(product.getId(), product.getTitle(), product.getImgURL(), product.getPrice());
     }
